Refactor: reimplement restructured producer display backend

+++ app/src/main/java/com/kssidll/arrugarq/data/repository/ProducerRepositorySource.kt
@@ -57,7 +57,7 @@ interface ProducerRepositorySource {
      * @param producer [ProductProducer] to get the total spending from
      * @return float representing total spending for the [producer] as flow
      */
-    fun totalSpentFlow(producer: ProductProducer): Flow<Float>
+    fun totalSpentFlow(producer: ProductProducer): Flow<Long>
 
     /**
      * @param producer [ProductProducer] to get the total spending by day from
